requirements.txt
aniso8601==1.2.0
asn1crypto==0.22.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
DateTime==4.2
dill==0.2.7.1
enum34==1.1.6
Flask==0.12.1
Flask-Ask==0.9.4
Flask-SQLAlchemy==2.2
future==0.16.0
futures==3.1.1
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
google-auth==1.2.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.7.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jws==0.1.3
MarkupSafe==1.0
nltk==3.2.5
oauth2client==3.0.0
paho-mqtt==1.3.0
pexpect==3.3
ply==3.8
pprint==0.1
protobuf==3.4.0
psycopg2==2.7.1
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.17
pycryptodome==3.4.3
PyJWT==1.5.3
pyOpenSSL==17.0.0
Pyrebase==3.0.27
python-dateutil==2.4.2
python-firebase==1.2
python-jwt==2.0.1
python-particle==0.2
pytz==2015.4
PyYAML==3.12
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
satori-sdk-python==1.0.3
six==1.10.0
SQLAlchemy==1.1.10
textblob==0.13.1
twilio==6.8.2
urllib3==1.22
Werkzeug==0.12.2
zope.interface==4.4.2
